# Common.Contracts/openapi_v3.yaml
openapi: 3.0.4
info:
  title: LPshows.live API v3
  description: This is the API for the Linkin Park Concert Calendar fan project
  version: 3.0.0
servers:
  - url: /
  - url: http://localhost:5133
paths:
...
  /v3/Tours:
    post:
      tags:
        - Tours
      summary: Creates a new tour
      operationId: CreateTour
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateTourRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateTourRequest"
        required: true
      responses:
        "200":
          description: ""
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/Tours
      security:
        - Bearer: []
    get:
      tags:
        - Tours
      summary: Returns a list of all tours
      operationId: GetTours
      responses:
        "200":
          description: List of all tours
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
            text/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/Tours
  /v3/Tours/{tourId}:
    get:
      tags:
        - Tours
      summary: Returns information about a tour
      operationId: GetTour
      parameters:
        - name: tourId
          in: path
          description: ID of the tour
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/Tour"
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
            text/json:
              schema:
                $ref: "#/components/schemas/Tour"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/tours/{tourId}
        requestParameters:
          integration.request.path.tourId: method.request.path.tourId
    delete:
      tags:
        - Tours
      summary: Deletes a tour
      operationId: DeleteTour
      parameters:
        - name: tourId
          in: path
          description: ID of the tour
          required: true
          schema:
            type: string
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: DELETE
        uri: https://api.dev.lpshows.felixsfd.de/v3/tours/{tourId}
        requestParameters:
          integration.request.path.tourId: method.request.path.tourId
      security:
        - Bearer: []
  /v3/Tours/{tourId}/legs:
    post:
      tags:
        - Tours
      summary: Creates a new leg of a tour
      operationId: CreateTourLeg
      parameters:
        - name: tourId
          in: path
          description: ID of the tour
          required: true
          schema:
            type: string
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTourLegRequest"
          text/json:
            schema:
              $ref: "#/components/schemas/AddTourLegRequest"
          application/*+json:
            schema:
              $ref: "#/components/schemas/AddTourLegRequest"
        required: true
      responses:
        "200":
          description: ""
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: POST
        uri: https://api.dev.lpshows.felixsfd.de/v3/tours/{tourId}/legs
        requestParameters:
          integration.request.path.tourId: method.request.path.tourId
      security:
        - Bearer: []
  /v3/Tours/{tourId}/legs/{legId}:
    get:
      tags:
        - Tours
      summary: Returns information about a tour leg
      operationId: GetTourLeg
      parameters:
        - name: tourId
          in: path
          description: ID of the tour
          required: true
          schema:
            type: string
        - name: legId
          in: path
          description: ID of the tour leg
          required: true
          schema:
            type: string
      responses:
        "200":
          description: information about the leg
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/TourLeg"
            application/json:
              schema:
                $ref: "#/components/schemas/TourLeg"
            text/json:
              schema:
                $ref: "#/components/schemas/TourLeg"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: GET
        uri: https://api.dev.lpshows.felixsfd.de/v3/tours/{tourId}/legs/{legId}
        requestParameters:
          integration.request.path.tourId: method.request.path.tourId
          integration.request.path.legId: method.request.path.legId
    delete:
      tags:
        - Tours
      summary: Deletes a leg of a tour
      operationId: DeleteTourLeg
      parameters:
        - name: tourId
          in: path
          description: ID of the tour
          required: true
          schema:
            type: string
        - name: legId
          in: path
          description: ID of the tour leg
          required: true
          schema:
            type: string
      responses:
        "200":
          description: no content
        "401":
          description: "Unauthorized: Credentials are missing or not valid"
        "403":
          description: "Forbidden: Credentials are valid, but the caller is not allowed to
            perform this operation"
      x-amazon-apigateway-integration:
        type: HTTP_PROXY
        httpMethod: DELETE
        uri: https://api.dev.lpshows.felixsfd.de/v3/tours/{tourId}/legs/{legId}
        requestParameters:
          integration.request.path.tourId: method.request.path.tourId
          integration.request.path.legId: method.request.path.legId
      security:
        - Bearer: []
...
components:
  schemas:
    AddTourLegRequest:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    CreateTourRequest:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    Tour:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        legs:
          type: array
          items:
            $ref: "#/components/schemas/TourLeg"
    TourLeg:
      required:
        - tourId
        - id
        - name
      type: object
      properties:
        tourId:
          type: string
        id:
          type: string
        name:
          type: string
...
  securitySchemes:
    Bearer:
      type: http
      description: JWT Bearer authentication
      scheme: Bearer
    ConcertsCognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
      x-amazon-apigateway-authorizer:
        type: cognito_user_pools
        providerARNs:
          - Fn::ImportValue:
              Fn::Sub: '${BaseProductName}-cognito-authorizer-user-pool'
...